--- app/src/main/java/com/vadimsalavatov/mobiledev/EmailConfirmationViewModel.kt
@@ -1,5 +1,0 @@
-package com.vadimsalavatov.mobiledev
-
-class EmailConfirmationViewModel : BaseViewModel() {
-
-}
