--- app/src/main/java/com/vadimsalavatov/mobiledev/MainFragmentViewModel.kt
@@ -1,5 +1,0 @@
-package com.vadimsalavatov.mobiledev
-
-class MainFragmentViewModel : BaseViewModel() {
-
-}
