--- app/src/main/java/com/vadimsalavatov/mobiledev/MainViewModel.kt
@@ -1,3 +1,0 @@
-package com.vadimsalavatov.mobiledev
-
-class MainViewModel : BaseViewModel()
