--- app/src/main/java/com/vadimsalavatov/mobiledev/SignInViewModel.kt
@@ -1,5 +1,0 @@
-package com.vadimsalavatov.mobiledev
-
-class SignInViewModel : BaseViewModel() {
-
-}
