--- app/src/main/java/com/vadimsalavatov/mobiledev/SignUpViewModel.kt
@@ -1,5 +1,0 @@
-package com.vadimsalavatov.mobiledev
-
-class SignUpViewModel : BaseViewModel() {
-
-}
