+++ app/src/main/java/com/vadimsalavatov/mobiledev/ui/MainActivity.kt
@@ -1,9 +1,10 @@
-package com.vadimsalavatov.mobiledev
+package com.vadimsalavatov.mobiledev.ui
 
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import com.vadimsalavatov.mobiledev.databinding.ActivityMainBinding
 import by.kirich1409.viewbindingdelegate.viewBinding
+import com.vadimsalavatov.mobiledev.R
 
 class MainActivity : AppCompatActivity(R.layout.activity_main) {
     private val viewModel: MainViewModel by viewModels()
