+++ app/src/main/java/com/vadimsalavatov/mobiledev/ui/MainViewModel.kt
@@ -1,0 +1,5 @@
+package com.vadimsalavatov.mobiledev.ui
+
+import com.vadimsalavatov.mobiledev.ui.base.BaseViewModel
+
+class MainViewModel : BaseViewModel()
