+++ app/src/main/java/com/vadimsalavatov/mobiledev/ui/base/BaseFragment.kt
@@ -1,8 +1,11 @@
-package com.vadimsalavatov.mobiledev
+package com.vadimsalavatov.mobiledev.ui.base
 
 import androidx.annotation.LayoutRes
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
+import com.vadimsalavatov.mobiledev.BuildConfig
+import com.vadimsalavatov.mobiledev.logBackstack
+import com.vadimsalavatov.mobiledev.logFragmentHierarchy
 import timber.log.Timber
 
 open class BaseFragment : Fragment {
