+++ app/src/main/java/com/vadimsalavatov/mobiledev/ui/base/BaseViewModel.kt
@@ -1,4 +1,4 @@
-package com.vadimsalavatov.mobiledev
+package com.vadimsalavatov.mobiledev.ui.base
 
 import androidx.lifecycle.ViewModel
