+++ app/src/main/java/com/vadimsalavatov/mobiledev/ui/emailconfirmation/EmailConfirmationFragment.kt
@@ -1,8 +1,10 @@
-package com.vadimsalavatov.mobiledev
+package com.vadimsalavatov.mobiledev.ui.emailconfirmation
 
 import androidx.fragment.app.viewModels
+import by.kirich1409.viewbindingdelegate.viewBinding
+import com.vadimsalavatov.mobiledev.R
 import com.vadimsalavatov.mobiledev.databinding.FragmentEmailConfirmationBinding
-import com.vadimsalavatov.mobiledev.databinding.FragmentSignInBinding
+import com.vadimsalavatov.mobiledev.ui.base.BaseFragment
 
 class EmailConfirmationFragment : BaseFragment(R.layout.fragment_email_confirmation) {
