+++ app/src/main/java/com/vadimsalavatov/mobiledev/ui/emailconfirmation/EmailConfirmationViewModel.kt
@@ -1,0 +1,7 @@
+package com.vadimsalavatov.mobiledev.ui.emailconfirmation
+
+import com.vadimsalavatov.mobiledev.ui.base.BaseViewModel
+
+class EmailConfirmationViewModel : BaseViewModel() {
+
+}
