+++ app/src/main/java/com/vadimsalavatov/mobiledev/ui/main/MainFragmentViewModel.kt
@@ -1,0 +1,7 @@
+package com.vadimsalavatov.mobiledev.ui.main
+
+import com.vadimsalavatov.mobiledev.ui.base.BaseViewModel
+
+class MainFragmentViewModel : BaseViewModel() {
+
+}
