+++ app/src/main/java/com/vadimsalavatov/mobiledev/ui/signin/SignInFragment.kt
@@ -1,7 +1,10 @@
-package com.vadimsalavatov.mobiledev
+package com.vadimsalavatov.mobiledev.ui.signin
 
 import androidx.fragment.app.viewModels
+import com.vadimsalavatov.mobiledev.ui.base.BaseFragment
+import com.vadimsalavatov.mobiledev.R
 import com.vadimsalavatov.mobiledev.databinding.FragmentSignInBinding
+import by.kirich1409.viewbindingdelegate.viewBinding
 
 class SignInFragment : BaseFragment(R.layout.fragment_sign_in) {
