+++ app/src/main/java/com/vadimsalavatov/mobiledev/ui/signin/SignInViewModel.kt
@@ -1,0 +1,7 @@
+package com.vadimsalavatov.mobiledev.ui.signin
+
+import com.vadimsalavatov.mobiledev.ui.base.BaseViewModel
+
+class SignInViewModel : BaseViewModel() {
+
+}
