+++ app/src/main/java/com/vadimsalavatov/mobiledev/ui/signup/SignUpViewModel.kt
@@ -1,0 +1,7 @@
+package com.vadimsalavatov.mobiledev.ui.signup
+
+import com.vadimsalavatov.mobiledev.ui.base.BaseViewModel
+
+class SignUpViewModel : BaseViewModel() {
+
+}
