+++ app/src/main/res/layout/activity_main.xml
@@ -2,7 +2,7 @@
 <androidx.fragment.app.FragmentContainerView
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/fragmentContainerView"
-    android:name="com.vadimsalavatov.mobiledev.UserListFragment"
+    android:name="com.vadimsalavatov.mobiledev.ui.main.MainFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 </androidx.fragment.app.FragmentContainerView>
